Added variables to main.css
Added primary and secondary theme color variables to main.css, then updated stylesheets to use them. This is to make it easier for users to just edit the colors at the start of the main.css file and not have to go through all of the CSS. There are still a few colors set not using variables in homepage.css and subpages.css.